+++ ajax/dropdownMassiveActionAddValidator.php
@@ -45,7 +45,7 @@ if (isset($_POST["validatortype"])) {
     switch ($_POST["validatortype"]) {
         case 'user':
             echo "<input type='hidden' name='groups_id' value=0 />";
-            User::dropdown(['name'   => 'users_id_validate',
+            User::dropdown(['name'   => 'users_id_validate', // TODO
                 'entity' => $_SESSION["glpiactive_entity"],
                 'right'  => ['validate_request', 'validate_incident']
             ]);
@@ -57,7 +57,7 @@ if (isset($_POST["validatortype"])) {
             break;
 
         case 'group':
-            echo "<input type='hidden' name='users_id_validate' value=0 />";
+            echo "<input type='hidden' name='users_id_validate' value=0 />"; // TODO
             $rand = Group::dropdown(['name'      => 'groups_id',
                 'entity'    => $_SESSION["glpiactive_entity"]
             ]);
@@ -107,14 +107,14 @@ if (isset($_POST["validatortype"])) {
             $param['display'] = true;
             $param['size']    = count($users);
 
-            Dropdown::showFromArray("users_id_validate", $users, $param);
+            Dropdown::showFromArray("users_id_validate", $users, $param); // TODO
 
            // Display all/none buttons to select all or no users in group
             if (!empty($_POST['groups_id'])) {
                 echo "<a id='all_users' class='btn btn-primary'>" . __('All') . "</a>";
                 $param_button = [
                     'validatortype'     => 'group_user',
-                    'users_id_validate' => '',
+                    'users_id_validate' => '', // TODO
                     'all_users'         => 1,
                     'groups_id'         => $_POST['groups_id'],
                     'right'             => ['validate_request', 'validate_incident'],