+++ ajax/dropdownValidator.php
@@ -42,66 +42,81 @@ header("Content-Type: text/html; charset=UTF-8");
 Html::header_nocache();
 
 if (isset($_POST["validatortype"])) {
-    if (isset($_POST['users_id_validate'])) {
-        Toolbox::deprecated('Usage of "users_id_validate" parameter is deprecated in "ajax/dropdownValidator.php". Use "items_id_target" instead.');
-        $_POST['items_id_target'] = $_POST['users_id_validate'];
+    if (array_key_exists('name', $_POST)) {
+        Toolbox::deprecated('Usage of "name" parameter is deprecated in "ajax/dropdownValidator.php". Use "prefix" instead.');
+        $itemtype_name      = 'itemtype_validate';
+        $items_id_name      = !empty($_POST['name']) ? $_POST['name'] . '[]' : 'users_id_validate[]';
+        $groups_id_name     = 'groups_id';
+    } else {
+        $itemtype_name      = $_POST['prefix'] . '[itemtype]';
+        $items_id_name      = $_POST['prefix'] . '[items_id]';
+        $groups_id_name     = $_POST['prefix'] . '[groups_id]';
     }
 
-    switch (strtolower($_POST["validatortype"])) {
+    if (array_key_exists('users_id_validate', $_POST)) {
+        Toolbox::deprecated('Usage of "users_id_validate" parameter is deprecated in "ajax/dropdownValidator.php". Use "items_id_target" instead.');
+        if (isset($_POST['users_id_validate']['groups_id'])) {
+            $_POST['groups_id'] = $_POST['users_id_validate']['groups_id'];
+        } else {
+            $_POST['itemtype'] = User::class;
+            $_POST['items_id'] = $_POST["validatortype"] !== 'list_users'
+                ? (isset($_POST['users_id_validate'][0]) ? $_POST['users_id_validate'][0] : 0)
+                : (is_array($_POST['users_id_validate']) ? $_POST['users_id_validate'] : []);
+        }
+    }
+
+    switch (strtolower($_POST['validatortype'])) {
         case 'user':
-            if (isset($_POST['items_id_target']['groups_id'])) {
-                $_POST['items_id_target'] = [];
-            }
-            $value = (isset($_POST['items_id_target'][0]) ? $_POST['items_id_target'][0] : 0);
             User::dropdown([
-                'name'   => !empty($_POST['name']) ? $_POST['name'] . '[]' : 'items_id_target[]',
+                'name'   => $items_id_name,
                 'entity' => $_POST['entity'],
-                'value'  => $value,
+                'value'  => $_POST['items_id'],
                 'right'  => $_POST['right'],
                 'width'  => '100%',
             ]);
-            echo Html::hidden('itemtype_target', ['value' => 'User']);
+            echo Html::hidden($itemtype_name, ['value' => 'User']);
             break;
 
         case 'group':
-            $value = (isset($_POST['items_id_target'][0]) ? $_POST['items_id_target'][0] : 0);
             Group::dropdown([
                 'name'   => !empty($_POST['name']) ? $_POST['name'] . '[]' : 'items_id_target[]',
                 'entity' => $_POST['entity'],
-                'value'  => $value,
+                'value'  => $_POST['items_id'],
                 'right'  => $_POST['right'],
                 'width'  => '100%',
             ]);
-            echo Html::hidden('itemtype_target', ['value' => 'Group']);
+            echo Html::hidden($itemtype_name, ['value' => 'Group']);
             break;
 
         case 'group_user':
-            $name = 'groups_id';
-            $value = $_POST['groups_id'];
+            $value = $_POST['groups_id'] ?? 0;
 
             $rand = Group::dropdown([
-                'name'   => $name,
+                'name'   => $groups_id_name,
                 'value'  => $value,
                 'entity' => $_POST["entity"],
                 'width'  => '100%',
             ]);
-            echo Html::hidden('itemtype_target', ['value' => 'User']);
+            echo Html::hidden($itemtype_name, ['value' => 'User']);
 
-            $param                        = ['validatortype' => 'list_users'];
-            $param['name']                = !empty($_POST['name']) ? $_POST['name'] : '';
-            $param['items_id_target']   = isset($_POST['items_id_target'])
-                                             ? $_POST['items_id_target'] : '';
-            $param['right']               = $_POST['right'];
-            $param['entity']              = $_POST["entity"];
-            $param['groups_id']           = '__VALUE__';
+            $param = [
+                'prefix'        => $_POST['prefix'],
+                'validatortype' => 'list_users',
+                'right'         => $_POST['right'],
+                'entity'        => $_POST['entity'],
+                'groups_id'     => '__VALUE__',
+            ];
+            if (array_key_exists('name', $_POST)) {
+                // TODO Drop in GLPI 11.0
+                $param['name'] = !empty($_POST['name']) ? $_POST['name'] : '';
+            }
             Ajax::updateItemOnSelectEvent(
-                "dropdown_$name$rand",
+                "dropdown_{$groups_id_name}{$rand}",
                 "show_list_users",
                 $CFG_GLPI["root_doc"] . "/ajax/dropdownValidator.php",
                 $param
             );
             if ($value) {
-                $param['validatortype'] = 'list_users';
                 $param['groups_id']     = $value;
                 Ajax::updateItem(
                     'show_list_users',
@@ -113,10 +128,8 @@ if (isset($_POST["validatortype"])) {
             break;
 
         case 'list_users':
-            if (isset($_POST['groups_id'])) {
-                $_POST['items_id_target'] = [];
-            }
-            $opt             = ['groups_id' => $_POST["groups_id"],
+            $opt             = [
+                'groups_id' => $_POST["groups_id"],
                 'right'     => $_POST['right'],
                 'entity'    => $_POST["entity"]
             ];
@@ -124,8 +137,8 @@ if (isset($_POST["validatortype"])) {
             $users           = [];
             $param['values'] = [];
             $values          = [];
-            if (isset($_POST['items_id_target']) && is_array($_POST['items_id_target'])) {
-                $values = $_POST['items_id_target'];
+            if (isset($_POST['items_id']) && is_array($_POST['items_id'])) {
+                $values = $_POST['items_id'];
             }
             foreach ($data_users as $data) {
                 $users[$data['id']] = formatUserName(
@@ -151,7 +164,7 @@ if (isset($_POST["validatortype"])) {
             $param['size']    = count($users);
 
             $rand  = Dropdown::showFromArray(
-                !empty($_POST['name']) ? $_POST['name'] : 'items_id_target',
+                $items_id_name,
                 $users,
                 $param
             );
@@ -160,7 +173,7 @@ if (isset($_POST["validatortype"])) {
             if (!empty($_POST['groups_id'])) {
                 $param_button = [
                     'validatortype'     => 'list_users',
-                    'name'              => !empty($_POST['name']) ? $_POST['name'] : '',
+                    'name'              => $items_id_name,
                     'items_id_target'   => '',
                     'all_users'         => 1,
                     'groups_id'         => $_POST['groups_id'],