+++ src/LevelAgreementLevel.php
@@ -221,11 +221,11 @@ abstract class LevelAgreementLevel extends RuleTicket
         unset($actions['olas_id']);
         unset($actions['slas_id']);
        // Could not be used as criteria
-        unset($actions['users_id_validate_requester_supervisor']);
-        unset($actions['users_id_validate_assign_supervisor']);
+        unset($actions['users_id_validate_requester_supervisor']); // TODO
+        unset($actions['users_id_validate_assign_supervisor']); // TODO
         unset($actions['affectobject']);
         unset($actions['groups_id_validate']);
-        unset($actions['users_id_validate']);
+        unset($actions['users_id_validate']); // TODO
         unset($actions['validation_percent']);
         $actions['status']['name']    = __('Status');
         $actions['status']['type']    = 'dropdown_status';